+++ lib/rubocop/cop/mixin/code_length.rb
@@ -14,7 +14,7 @@ module RuboCop
         cop_config['CountComments']
       end
 
-      def check_code_length(node, *_)
+      def check_code_length(node)
         length = code_length(node)
         return unless length > max_length